update sell funtion

+++ programs/mushi_program/src/context/common.rs
@@ -96,37 +96,37 @@ impl<'info> ACommon<'info> {
     }
     pub fn sol_to_mushi(&self, sol_amount: u64) -> Result<u64>{
         Ok(
-            ((sol_amount as u128).checked_mul(self.global_state.token_supply as u128).unwrap() as u64)
+            ((sol_amount as u128).checked_mul(self.global_state.token_supply as u128).unwrap())
             .checked_div(
-                self.get_backing()?.checked_sub(sol_amount).unwrap()
-            ).unwrap()
+                self.get_backing()?.checked_sub(sol_amount).unwrap() as u128
+            ).unwrap() as u64
         )
     }
     pub fn mushi_to_sol(&self, mushi_amount: u64) -> Result<u64>{
         Ok(
-            mushi_amount.checked_mul(self.get_backing()?).unwrap()
-            .checked_div(self.global_state.token_supply).unwrap()
+            ((mushi_amount as u128).checked_mul(self.get_backing()? as u128).unwrap())
+            .checked_div(self.global_state.token_supply as u128).unwrap() as u64
         )
     }
     pub fn sol_to_mushi_lev(&self, sol_amount: u64, fee: u64) -> Result<u64>{
         let backing = self.get_backing()? - fee;
         Ok(
-            sol_amount.checked_mul(self.global_state.token_supply).unwrap()
-            .checked_add(backing - 1).unwrap()
-            .checked_div(backing).unwrap()
+            ((sol_amount as u128).checked_mul(self.global_state.token_supply as u128).unwrap())
+            .checked_add((backing - 1) as u128).unwrap()
+            .checked_div(backing as u128).unwrap() as u64
         )
     }
     pub fn sol_to_mushi_no_trade_ceil(&self, sol_amount: u64) -> Result<u64>{
         Ok(
-            sol_amount.checked_mul(self.global_state.token_supply).unwrap()
-            .checked_add(self.get_backing()? - 1).unwrap()
-            .checked_div(self.get_backing()?).unwrap()
+            ((sol_amount as u128).checked_mul(self.global_state.token_supply as u128).unwrap())
+            .checked_add(self.get_backing()? as u128 - 1).unwrap()
+            .checked_div(self.get_backing()? as u128).unwrap() as u64
         )
     }
     pub fn sol_to_mushi_no_trade(&self, sol_amount: u64) -> Result<u64>{
         Ok(
-            sol_amount.checked_mul(self.global_state.token_supply).unwrap()
-            .checked_div(self.get_backing()?).unwrap()
+            ((sol_amount as u128).checked_mul(self.global_state.token_supply as u128).unwrap())
+            .checked_div(self.get_backing()? as u128).unwrap() as u64
         )
     }
     pub fn safety_check(&mut self) -> Result<()> {
